🫡 First commit - making things work cross-platform
Many things were done here.  First was the addition of dev containers,
allowing for more consistent and easy development across all platforms.
(Especially us on OSX.)

The csproj was greatly simplified to work with newer dotnet build tools,
and should still build just fine in full Visual Studio.  Updating the
sln will follow.

Otherwise things just mostly-worked once all the references were fixed.
A script for pulling the Rimworld-specific dependencies was added, but
that is OSX-specific for the moment and only needed to make Intellisense
work.  Another for local publishing was added as well.

This and more are explained in the added README.