🎨 Gradient coloring added in the UI
Selection of style remains to be done still, and an improved UI is needed
now that anything is selectable.  Most importantly is a gradient style
selector, but also a color palette that provides options for gradients
along the current character's selection, and sensible defaults.

Using the new color selectors also hits beard colors simulatiously, but
in testing this looks like a pre-existing bug.